requirements.txt
websockets==12.0
pydantic==2.6.4
pydantic-settings==2.2.1
structlog==24.1.0
rich==13.7.0
httpx==0.27.0
python-dotenv==1.0.1
